refactor(daemon): organize server classes into directories
