+++ test/daemon/Tiempitod.Tests/Sessions/SessionManagerTests.cs
@@ -5,7 +5,7 @@ using Tiempitod.NET;
 using Tiempitod.NET.Configuration.Notifications;
 using Tiempitod.NET.Configuration.Session;
 using Tiempitod.NET.Notifications;
-using Tiempitod.NET.Server;
+using Tiempitod.NET.Server.StandardOut;
 using Tiempitod.NET.Session;
 using Tiempitod.Tests.Sessions.Helpers;
